a draft of the guarantees
This is a draft of the guarantees; this is in some ways the 'point' of it all, why a government should even exist. we start in order of Maslow's needs, then move into some somewhat more abstract ones, which are also more likely to be poorly phrased and need adjusting. 

Also, things are generally phrased as "have access to" in order to try to avoid governmental over-command; people need choice, the government shouldn't mandate these things too strongly, but it should do what it takes to make it possible for people to be able to choose good things.