Define citizenship in the polity
Citizenship needs definition, and protection.

This attempts to define citizenship for a potentially-new polity in a reasonable way, embracing the US's birthright citizenship in a way that will be harder to attempt to undo as some people have been trying to do in the US.

The definition of what's required to revoke citizenship is structured to avoid ever putting someone in a legal limbo - if they actually want to leave and not be a citizen anymore they can as long as it doesn't render them stateless, otherwise no, you cannot.

Naturalization is permitted but not defined, denaturalization is not except via above. "false naturalization" could be created as a crime with punishments but those punishments will not allow the revocation of citizenship because that potentially allows people tools to abuse.

The expansion clause may well be moot, since expansions will probably come with legislation and/or amendments regularizing what happens anyway, but why not make it a little harder to be an extractive colonial state?