define electoral oversight
The electoral oversight body is a small, largely procedural body, for the purpose of overseeing sortition and plebiscites; it's my solution to the problem of "he who counts the votes".

We select it in a very specific process that has very limited discretion available, which should be easily verified, but also is probably going to be kind of slow and shitty and annoying, so they get to have other procedures for sortitions that aren't for positions that count the votes. You can also never hold an EO position twice, because even if that happened by pure chance it is the kind of thing that would just make so many conspiracy theories happen, and ... well, if someone figures out how to make it happen, we definitely don't want to let them.

Other than that, we also include some rules for what sortition implies for the polity as a whole; minimum salary for sortitioned positions to reduce the threat of bribery and corruption, the right of refusal because if people aren't willing to do it, we can't force them, etc.